In the case of a statement that provides a column label, use the label
The RDS response includes both the field `Name` and `Label` as an
indication of the original column name and the label assigned by the query.
In the case where there is no label they are identical, however
in order to map the correct column name when a label is used, we choose
to use the Label field instead.